+++ addOns/websocket/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Maintenance changes.
+- Use the Requester add-on for the WebSocket Message Editor dialogues.
 
 ### Fixed
 - Correctly handle no message in the WebSocket message panel.